refactor: db 힐트 모듈 일원화
